Update power trace script and add IDs script (#74)
The power trace script needed to be moved out of the fp16 folder. It
also needs a couple extra steps done to keep up with some new issues in
smutrace.

The gpu_ids.py script is a clumsy way to relate the divergent IDs that
GPUs get in various tools. No warranty or guarantee!